requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-celery-beat==2.4.0
django-crontab==0.7.1
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.13.1
drf-yasg==1.21.3
exceptiongroup==1.1.0
flower==1.2.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.4.0
idna==3.3
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
pika==0.13.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytest==7.2.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-telegram-bot==20.0
pytz==2022.2.1
redis==4.4.2
requests==2.28.1
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.6
